requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
httpx
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
bcrypt==4.0.1
jinja2
websockets
tzdata>=2023.3